+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 * server: map FileNotFoundException and NoSuchFileException to 404 status code by default
+* core: TSIDGenerator.generateValue() can now accept desired timestamp instead of current one
 
 # 1.8.0
 * server: StatusCodeExceptions will now contain stack trace in dev mode