+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
+* JDK 17+ is now required
 * jdbc: support for null conversion of @JvmInline classes (e.g. TSID) when reading from DB
 * server: detectBrowser() will not detect bots in addition to the browser version they emulate